Added full email element in the json input string;

+++ update_user_info.sql
@@ -17,6 +17,7 @@ DECLARE
   o_message text;
   i_business_unit text;
   i_email text;
+  i_full_email text;
   i_full_name text;
   i_status text;
   i_steward_group json;
@@ -31,6 +32,7 @@ BEGIN
   SELECT $1::json->>'user_id',
          $1::json->>'business_unit',
 		 $1::json->>'email',
+		 $1::json->>'fullemail',
 		 $1::json->>'fullname',
 		 $1::json->>'status',
 		 $1::json->'steward_group',
@@ -38,6 +40,7 @@ BEGIN
     INTO o_userid,
          i_business_unit,
 		 i_email,
+		 i_full_email,
 		 i_full_name,
 		 i_status,
 		 i_steward_group,
@@ -52,11 +55,15 @@ BEGIN
   RAISE NOTICE 'BU %', i_business_unit;
   RAISE NOTICE 'UID %', o_userid;
   RAISE NOTICE 'email %', i_email;
+  RAISE NOTICE 'full email %', i_full_email;
   RAISE NOTICE 'fullName %', i_full_name;
   RAISE NOTICE 'Status %', i_status;
   RAISE NOTICE 'GRP %', i_steward_group;
   RAISE NOTICE 'UName %', i_user_name; 
   --BEGIN TRANSACTION;
+  IF NOT is_blank_string(i_full_email) THEN 
+    i_email = i_full_email;
+  END IF;
   IF isnew THEN       
     INSERT INTO ige_user (username, email, business_unit, status, status_date, full_name)	            
 	  VALUES 	  
@@ -70,9 +77,10 @@ BEGIN
 	  )
 	  RETURNING user_id::text INTO o_userid;
 	  -- Beginning of updating Oracle to sync
-	  INSERT INTO imaint_oracle.ige_user(user_id, username, email, business_unit, status, status_date, first_name, last_name)
-	  VALUES 	  
-	  (
+	  IF get_configuration_bool('anchorTO', 'ANCHORTO', 'sync_with_oracle') THEN
+	    INSERT INTO imaint_oracle.ige_user(user_id, username, email, business_unit, status, status_date, first_name, last_name)
+	    VALUES 	  
+	    (
 		  o_userid::int,
 		  i_user_name,
 		  i_email,
@@ -81,9 +89,10 @@ BEGIN
 		  now(),
 		  split_part(i_full_name, ' ', 1),
 		  split_part(i_full_name, ' ', 2)
-	  );
-	  -- End of updating Oracle to sync
-	  RAISE NOTICE 'Updated Oracle ige_user table: %', o_userid;
+	    );
+		RAISE NOTICE 'Updated Oracle ige_user table: %', o_userid;
+	  END IF;
+	  -- End of updating Oracle to sync	  
   ELSE    
     UPDATE ige_user
 	  SET username = i_user_name,
@@ -97,18 +106,20 @@ BEGIN
 	DELETE FROM ige_user_steward 
 	WHERE user_id = o_userid::bigint;
 	-- Beginning of updating Oracle to sync
-	UPDATE imaint_oracle.ige_user
-	  SET username = i_user_name,
+	IF get_configuration_bool('anchorTO', 'ANCHORTO', 'sync_with_oracle') THEN
+	  UPDATE imaint_oracle.ige_user
+	    SET username = i_user_name,
 		  email = i_email,
 		  business_unit = i_business_unit,
 		  status = i_status,
 		  status_date = now(),
 		  first_name = split_part(i_full_name, ' ', 1),
 		  last_name = split_part(i_full_name, ' ', 1)
-	WHERE user_id = o_userid::bigint;
-	-- Remove old user stewards
-	DELETE FROM imaint_oracle.ige_user_steward 
-	WHERE user_id = o_userid::bigint;
+	  WHERE user_id = o_userid::bigint;
+	  -- Remove old user stewards
+	  DELETE FROM imaint_oracle.ige_user_steward 
+	  WHERE user_id = o_userid::bigint;
+	END IF;  
 	-- End of updating Oracle to sync
   END IF;
   
@@ -120,7 +131,8 @@ BEGIN
 	--FROM json_array_elements_text(i_steward_group);
 	
   -- Beginning of updating Oracle to sync
-  INSERT INTO imaint_oracle.ige_user_steward
+  IF get_configuration_bool('anchorTO', 'ANCHORTO', 'sync_with_oracle') THEN
+    INSERT INTO imaint_oracle.ige_user_steward
     SELECT user_id,
 	       steward_group,
 		   'N',
@@ -128,7 +140,8 @@ BEGIN
 	FROM ige_user_steward   
     WHERE user_id = o_userid::int; 
 --	       json_array_elements_text(i_steward_group);   
-  RAISE NOTICE 'Updated Oracle ige_user_steward table: %', o_userid; 
+    RAISE NOTICE 'Updated Oracle ige_user_steward table: %', o_userid; 
+  END IF;
   -- End of updating Oracle to sync
   SELECT row_to_json(c) INTO o_json
 	FROM
@@ -153,7 +166,7 @@ BEGIN
 	RETURN o_json;
   WHEN 	foreign_key_violation THEN
     o_status = SQLSTATE;
-	o_message = 'Invalid status';	
+	o_message = 'Invalid status: Foreign key violation';	
 	SELECT row_to_json(c) INTO o_json
 	FROM
 	(
